Add Embedded Compound Extension
This patch adds the ability to be embedded as a Compound Extension. To do so, we simply need to use a wrapper for our Web3 provider, `RPCWeb3Provider`, which communicates via iframe communication. Additionally, once we detect we're loading in an embedded state (via `?embedded` search param), we don't show the top-nav header bar. Additionally, we assume we can auto-connect to a Web3 account, since the connection logic is handled by the Compound III app, itself.

Finally, this patch includes the ability to vote on all proposals, even if they are not in voting states. This is just for testing, but should be removed before merging.